+++ packages/dc43-service-clients/src/dc43_service_clients/data_products/__init__.py
@@ -1,5 +1,6 @@
 """Client-facing data product service helpers."""
 
+from ._compat import DataProductRegistrationResult
 from .client.interface import DataProductServiceClient
 from .client.local import LocalDataProductServiceClient
 from .client.remote import RemoteDataProductServiceClient
@@ -9,7 +10,6 @@ from .models import (
     normalise_input_binding,
     normalise_output_binding,
 )
-from dc43_service_backends.data_products import DataProductRegistrationResult
 
 __all__ = [
     "DataProductInputBinding",